requirements.txt
asyncio
pyrogram
python-dotenv
